Add files via upload
In this dataset, we'll be analyzing a little over 6k Netflix shows focusing on the day that Netflix upload most of it's shows

There will be several factors to consider such as days of the week and others. The goal is to determine the day that Netflix upload most of it's shows.